NODE-1104: Moved the BlockHash and DeployHash type aliases.
