+++ storage/src/main/scala/io/casperlabs/storage/util/TopologicalSortUtil.scala
@@ -3,7 +3,7 @@ package io.casperlabs.storage.util
 import com.google.protobuf.ByteString
 import io.casperlabs.casper.consensus.Block
 import io.casperlabs.crypto.codec.Base16
-import io.casperlabs.storage.block.BlockStorage.BlockHash
+import io.casperlabs.storage.BlockHash
 
 object TopologicalSortUtil {
   type BlockSort = Vector[Vector[BlockHash]]
