+++ storage/src/test/scala/io/casperlabs/storage/deploy/MockDeployStorage.scala
@@ -9,7 +9,7 @@ import io.casperlabs.casper.consensus.info.DeployInfo
 import io.casperlabs.crypto.codec.Base16
 import io.casperlabs.crypto.Keys.{PublicKey, PublicKeyBS}
 import io.casperlabs.shared.Log
-import io.casperlabs.storage.block.BlockStorage.{BlockHash, DeployHash}
+import io.casperlabs.storage.{BlockHash, DeployHash}
 import io.casperlabs.storage.deploy.MockDeployStorage.Metadata
 import io.casperlabs.shared.Sorting.byteStringOrdering
